standalone_interface/requirements.txt
# Standalone Watus Interface Requirements
Flask>=3.0.0
Flask-SocketIO>=5.0.0
pyzmq>=25.0.0
python-dotenv>=1.0.0
eventlet>=0.35.0
